Add allowlist filtering (#1)
Previously tables that are to be excluded are listed explicitly. This allows the opposite logic; tables to be included can be listed explicitly. In this way all tables can have by default `ignore` or `nodata`, and any other value (eg `include`) will include them. See the documentation for more information.